--- data/duke.txt
@@ -1,1 +1,0 @@
-T,false,shit
