+++ src/main/java/duke/storage/CSVParsable.java
@@ -1,0 +1,5 @@
+package duke.storage;
+
+public interface CSVParsable {
+	public CSV toCSV();
+}
