+++ src/main/java/duke/task/TaskType.java
@@ -1,0 +1,5 @@
+package duke.task;
+
+enum TaskType {
+    TODO_TASK, DEADLINE_TASK, EVENT_TASK
+}
